Move more scanners to the right module
